requirements.txt
astroid==2.9.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
coverage==6.2
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
freezegun==1.1.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
Inject==4.3.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.1
pluggy==1.0.0
psycopg2==2.9.2
py==1.11.0
pycodestyle==2.8.0
pycryptodome==3.12.0
pyflakes==2.4.0
pylint==2.12.2
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.6.1
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.29
tinytuya==1.3.0
toml==0.10.2
tomli==2.0.0
typing-extensions==4.0.1
urllib3==1.26.7
w1thermsensor==2.0.0
Werkzeug==2.0.2
wrapt==1.13.3

celery~=5.2.3
setuptools~=59.6.0